chore: move UpsertKV to separate file
